// lib/schemas/userSchema.ts
import { z } from 'zod';

export const userSchema = z.object({
  name: z.string().min(1, { message: 'Name is required' }),
  email: z.string().email({ message: 'Invalid email address' }),
  password: z.string().min(6, { message: 'Password must be at least 6 characters' }),
  role: z.enum(['admin', 'wholesale1', 'wholesale2', 'retail'], { message: 'Invalid role' }),
});
export type UserFormData = z.infer<typeof userSchema>;